Fix: Fix the persian font + some changes - Add the persian name with Vazir font - Dynamically change the fontsize of the description (the display is still too small) - Minor changes
